Added Text type
This is the main reason for rewriting Value: thanks to the Proxy, we get
naturally the Array like access to string characters. We do this by
passing a handler with a special getter to the Value factory that
can handle the case where a number is called as property key.